requirements.txt
pycryptodome~=3.23.0
matplotlib~=3.10.8
numpy~=2.4.1